discussion_show/requirements.txt
langchain-openai
langchain-community
langchain
python-dotenv
runpod
nicegui
openai
pydub
requests
